Add a script to add links to KA to all our videos
Summary: I need to update the site code to not ever store the link, but apart from that I think this is ready.

Test Plan:
Ran the test file.

Checked that `youtube_ids` contains 3000+ things that look like youtube IDs.

Overrode `youtube_ids` to contain a video in my uploads; ran the script and it added the link. Ran the script again and skipped over the video because it already had the link.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D773